Ref: Created /create_card and moved file inside
